Fixed crashes and use RAII recursive mutex locking, audio is still broken as I don't know what I'm doing with the conversion from float to u16 :)

+++ include/oddlib/audio/Sample.h
@@ -8,5 +8,5 @@ public:
     Uint16 * m_SampleBuffer;
     unsigned int i_SampleSize;
 
-    float GetSample(float sampleOffset, bool interpolation);
+    Uint16 GetSample(Uint32 sampleOffset, bool interpolation);
 };
